@@ -1,31 +1,30 @@
 ﻿using System;
 using System.Windows.Input;
 
-namespace IrregularVerbTrainer.Commands
+namespace IrregularVerbTrainer.Commands;
+
+public abstract class Command : ICommand
 {
-    public abstract class Command : ICommand
+    public event EventHandler CanExecuteChanged
     {
-        public event EventHandler CanExecuteChanged
-        {
-            add => CommandManager.RequerySuggested += value;
-            remove => CommandManager.RequerySuggested -= value;
-        }
-
-        public bool CanExecute(object parameter)
-        {
-            return CanExecuteCmd(parameter);
-        }
-
-        public void Execute(object parameter)
-        {
-            ExecuteCmd(parameter);
-        }
-
-        protected virtual bool CanExecuteCmd(object parameter)
-        {
-            return true;
-        }
-
-        protected abstract void ExecuteCmd(object parameter);
+        add => CommandManager.RequerySuggested += value;
+        remove => CommandManager.RequerySuggested -= value;
     }
+
+    public bool CanExecute(object parameter)
+    {
+        return CanExecuteCmd(parameter);
+    }
+
+    public void Execute(object parameter)
+    {
+        ExecuteCmd(parameter);
+    }
+
+    protected virtual bool CanExecuteCmd(object parameter)
+    {
+        return true;
+    }
+
+    protected abstract void ExecuteCmd(object parameter);
 }
